+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix bug where the difficulty of game variants may not be choosed even thought the difficulty is known of the game variant.
+- Fix bug where a restart of a game does not clear the possible numbers.
 
 ### Security